# byclaw-tenant-node/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ByClaw fixed tenant node",
    "version": "1.0.0",
    "description": "BE-only mTLS API. Canonical business hash and Redis/Worker contracts are specified in README. No raw SQL/bootstrap/message overwrite API."
  },
  "servers": [
    {
      "url": "https://tenant-node.example.test:3100"
    }
  ],
  "security": [
    {
      "mtls": []
    }
  ],
  "paths": {
...
    "/internal/v1/sessions/{id}": {
      "patch": {
        "summary": "UPDATE_SESSION",
        "parameters": [
          {
            "$ref": "#/components/parameters/enterprise"
          },
          {
            "$ref": "#/components/parameters/generation"
          },
          {
            "$ref": "#/components/parameters/actor"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[1-9][0-9]{0,18}$",
              "description": "规范十进制 signed bigint ID；禁止 JS Number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantCommand"
              }
            }
          }
        }
      },
      "delete": {
        "summary": "DELETE_SESSION",
        "parameters": [
          {
            "$ref": "#/components/parameters/enterprise"
          },
          {
            "$ref": "#/components/parameters/generation"
          },
          {
            "$ref": "#/components/parameters/actor"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[1-9][0-9]{0,18}$",
              "description": "规范十进制 signed bigint ID；禁止 JS Number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantCommand"
              }
            }
          }
        }
      },
      "get": {
        "summary": "Accessible session",
        "parameters": [
          {
            "$ref": "#/components/parameters/enterprise"
          },
          {
            "$ref": "#/components/parameters/generation"
          },
          {
            "$ref": "#/components/parameters/actor"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[1-9][0-9]{0,18}$",
              "description": "规范十进制 signed bigint ID；禁止 JS Number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/internal/v1/sessions/{id}/messages": {
      "get": {
        "summary": "Traditional session messages",
        "parameters": [
          {
            "$ref": "#/components/parameters/enterprise"
          },
          {
            "$ref": "#/components/parameters/generation"
          },
          {
            "$ref": "#/components/parameters/actor"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[1-9][0-9]{0,18}$",
              "description": "规范十进制 signed bigint ID；禁止 JS Number"
            }
          },
          {
            "name": "pageNum",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100000
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/internal/v1/sessions/query": {
      "post": {
        "summary": "Personal session list",
        "parameters": [
          {
            "$ref": "#/components/parameters/enterprise"
          },
          {
            "$ref": "#/components/parameters/generation"
          },
          {
            "$ref": "#/components/parameters/actor"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "mtls": {
        "type": "mutualTLS",
        "description": "Trusted BE certificate chain and pinned CN"
      }
    },
    "parameters": {
      "enterprise": {
        "name": "X-Enterprise-Id",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "^[1-9][0-9]{0,18}$",
          "description": "规范十进制 signed bigint ID；禁止 JS Number"
        }
      },
      "generation": {
        "name": "X-Tenant-Generation",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "^[1-9][0-9]{0,18}$",
          "description": "规范十进制 signed bigint ID；禁止 JS Number"
        }
      },
      "actor": {
        "name": "X-Actor-User-Id",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "^[1-9][0-9]{0,18}$",
          "description": "规范十进制 signed bigint ID；禁止 JS Number"
        }
      },
...
    },
    "schemas": {
      "TenantCommand": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "protocolVersion",
          "enterpriseId",
          "generation",
          "dbSandboxRecordId",
          "userId",
          "sessionId",
          "requestId",
          "operation",
          "tenantMemberUserIds",
          "requestHash",
          "payload"
        ],
        "properties": {
          "protocolVersion": {
            "const": 1
          },
          "enterpriseId": {
            "type": "string",
            "pattern": "^[1-9][0-9]{0,18}$",
            "description": "规范十进制 signed bigint ID；禁止 JS Number"
          },
          "generation": {
            "type": "string",
            "pattern": "^[1-9][0-9]{0,18}$",
            "description": "规范十进制 signed bigint ID；禁止 JS Number"
          },
          "dbSandboxRecordId": {
            "type": "string",
            "pattern": "^[1-9][0-9]{0,18}$",
            "description": "规范十进制 signed bigint ID；禁止 JS Number"
          },
          "userId": {
            "type": "string",
            "pattern": "^[1-9][0-9]{0,18}$",
            "description": "规范十进制 signed bigint ID；禁止 JS Number"
          },
          "sessionId": {
            "type": "string",
            "pattern": "^[1-9][0-9]{0,18}$",
            "description": "规范十进制 signed bigint ID；禁止 JS Number"
          },
          "requestId": {
            "type": "string",
            "pattern": "^[A-Za-z0-9:_-]{1,64}$"
          },
          "operation": {
            "type": "string",
            "enum": [
              "CREATE_SESSION",
              "CREATE_GROUP",
              "UPDATE_SESSION",
              "DELETE_SESSION",
              "ADD_MEMBERS",
              "JOIN_GROUP",
              "REMOVE_MEMBER",
              "SET_ROLE",
              "TRANSFER_OWNER",
              "LEAVE_GROUP",
              "DISSOLVE_GROUP",
              "ACK_DISSOLUTION",
              "UPDATE_SETTINGS",
              "READ_STATE",
              "CREATE_TASK",
              "UPDATE_TASK",
              "PUBLISH_TASK",
              "SAVE_PENDING_PUBLICATION",
              "DELETE_PENDING_PUBLICATION",
              "RECALL_MESSAGE"
            ]
          },
          "tenantMemberUserIds": {
            "type": "array",
            "items": {
              "type": "string",
              "pattern": "^[1-9][0-9]{0,18}$",
              "description": "规范十进制 signed bigint ID；禁止 JS Number"
            },
            "minItems": 1,
            "maxItems": 1000,
            "description": "仅由 BE 验证 ACTIVE 成员后生成"
          },
          "requestHash": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$"
          },
          "payload": {
            "type": "object",
            "description": "操作字段见 README；resourceAuthorized/joinLinkAuthorized 仅由 BE 生成"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code"],
            "properties": {
              "code": {
                "type": "string"
              }
            }
          }
        }
      },
...
    }
  }
}